requirements.txt
django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
